Don't store data on the dom. Remove getters. (d3.alt)
I removed the internal flat storage of the nodes; now there is only the
two-dimensional groups array, which can be inspected from the outside. For
ease of implementation, I removed the old getters that allowed you to query
attributes (etc.) since it doesn't make as much sense to do that when the
selection is grouped. And anyway, that was primarily intended for debugging
purposes. You can still use `each` for iteration.