Don't reselect on remove.
The previous behavior of selecting the parent node was extremely confusing
because it could result in the parent node being selected multiple times (when
removing siblings). Even worse, the child data would override the parent data!
This changes the `remove` operator so that it matches the behavior of the W3C
API (removeChild), returning the removed nodes.