Fix simulation restarting.
If the simulation cools and stops, we need to restart it, rather than just
resetting the alpha parameter. So, it is useful to have both `resume` and
`start` methods on the force layout. This commit also fixes a bug where
mousedown events on draggable nodes would continue to propagate, interfering
with other events (such as clicking to add a node).