Rewrite stack layout for flexibility.
The stack layout can now operate on arbitrary data structures, rather than
requiring a specific input format. To do this, you can now specify functions to
access values per series, along with x- and y-coordinates.

One tricky thing is that the stack layout needs to be able to write the stacked
values back into the data; this is accomplished using the customizable `out`
function. This function takes three arguments: the current datum (d), the offset
value (y0), and the y value. The y value is unchanged unless the "expand" offset
is used.

This commit also allows you to specify arbitrary functions for order and offset
computation, in addition to the built-in methods supported previously.