Add configurable origin to d3.behavior.zoom.
This defaults to storing the absolute scale and transform across
multiple zooms.  However, it can be modified to retrieve state from
outside the behaviour e.g. if the offsets have been restricted in some
way, the true values should be returned in a custom offset accessor.

The offset accessor is called on "mousedown" or "touchstart", and so has
access to the current mouse position.