Less zealous click cancelling.
The click event is prevented from propagating when a drag has occurred
i.e. the mouse has moved between mousedown and mouseup.  This is because
a click event will be triggered after a mousedown and mouseup have
occurred on the same element.  However, we were cancelling the next
click event *anywhere*, regardless of whether mousedown and mouseup
occurred on the same element or not.

As a sanity check, we also check that the click event was triggered on
the same element that mousedown *and* mouseup occurred on, before
cancelling it.  Just in case!