Add drag.origin accessor.
This allows you to specify how the drag behavior should infer the current
position of the targetted element, such that a fixed offset can be maintained
between the mouse position and the element position for the duration of the drag
gesture. For example, if your data is an {x, y} object, use origin(Object).