Allow decorative brush resizers.
The brush component now allows you to add decorative elements to serve as brush
resizers, while keeping the default behavior of a few invisible pixels on the
edge of the extent. Any element you add to ".resize" containers can serve as
visible handles for the brush.

This commit also fixes two small bugs with the brush component. First, we now
track the offset between the extent edge and the mouse, rather than immediately
jumping the extent edge to the mouse position; this was barely noticeable
previously due to the resizers being so thin, but it is very noticeable with
large resizers. Second, the default behavior for the SPACE key is not prevented
whilst dragging.