Inheritance for transition.each.
If you use d3.transition(selection) within the context of a transition.each, the
created transition will automatically inherit the delay, duration and easing of
the parent transition. Likewise, if you use d3.transition(selection) within the
context of a selection.each, it simply returns the passed-in selection with no
transition at all. This makes it easier to create components, such as the axis,
that support automatic transitions.