Add example for usage of `defined` function
The `defined` function accepts a constant that is evaluated as
truthy or falsey, or a function that returns such truthy or falsey
value. If the result is truthy, then it means that datum is defined. If
falsey, then that datum is considered undefined and will not be
represented visually. This is useful in drawing multi-segmented line
or area charts, using just one SVG `<path>` directive.