Expose access to request.responseType in xhr
Also, don't always prefer XDomainRequest in a cross domain request.

This allows the caller to specify the response value type where available,
such as blob or json, which can greatly accelerate dealing with certain
types of data.

A cursory search suggests that IE's XDomainRequest does not use responseType,
so the caller will need to check the resulting type if a specific
responseType is required for a cross domain request.

IE's XDomainRequest also doesn't appear to have many other aspects available
in XMLHttpRequest, such as responseXML, currently used in d3.xml.

So a feature test is done on XMLHttpRequest to see if it supports CORS,
rather than always preferring IE's XDomainRequest where a cross domain
request is needed.