Add pack.radius for explicit radius.
Fixes #857. By default, the radius of each circle in the pack layout is derived
from the square root of the node’s value, which is then scaled according to the
bounding box of all circles and the layout’s size. By specifying a radius
function, you can now explictly control the radius, disabling automatic scaling.

The radius function takes the value as input and returns the corresponding
radius. If no radius function is specified, automatic scaling is enabled; this
is the default. Conveniently, d3.scale.sqrt can be used as a radius function.