Use linear segments for basis interpolation ends.
The first and last segment of a line interpolated using basis (B-spline)
interpolation are necessarily linear. WebKit appears to have a bug where linear
Bézier segments do not compute tangents correctly (presumably because the
control points are coincident with an end point, and the angle of a zero-length
vector is undefined). We can workaround this browser bug by using linear
segments for the start and end.