Fix for mousewheel interrupting zoom transition.
If a zoom transition is in progress, using zoom.event, a mousewheel
gesture will interrupt the transition and cause a zoomend event, which
resets the center0 variable to null.  Previously, the center0 variable
was assigned in the mousewheel handler prior to the interruption.  This
fix assigns it after the interruption.